[refactor] 파일 이름 convention에 맞게 변경 - 기본적으로 kebab-case를 따른다 - class, type, interface, component의 이름은 PascalCase를 따른다
