+++ src/core/router/index.ts
@@ -10,7 +10,7 @@ const router = createRouter({
       // this generates a separate chunk (About.[hash].js) for this route
       // which is lazy-loaded when the route is visited.
       // component: () => import("../views/StuffsView.vue")
-      component: () => import('@^stuffs/StuffsPage.vue'),
+      component: () => import('@^stuffs/StuffPage.vue'),
     },
     {
       path: '/histories',
@@ -18,7 +18,7 @@ const router = createRouter({
       // route level code-splitting
       // this generates a separate chunk (About.[hash].js) for this route
       // which is lazy-loaded when the route is visited.
-      component: () => import('@^histories/HistoriesPage.vue'),
+      component: () => import('@^histories/HistoryPage.vue'),
     },
   ],
 });
