+++ src/modules/histories/components/HistoryPageOnCommon/HistoryPageOnCommon.vue
@@ -4,8 +4,8 @@ import { computed } from 'vue';
 
 import HistoryDetail from '@^histories/components/HistoryDetailSection/HistoryDetailSection.vue';
 import HistoryList from '@^histories/components/HistoryList/HistoryList.vue';
-import { getHistoryListQuery } from '@^histories/components/utils/utils';
-import useHistorySelectedStore from '@^histories/stores/historySelectedStore';
+import { getHistoryListQuery } from '@^histories/components/utils/history-query-utils';
+import useHistorySelectedStore from '@^histories/stores/history-selected-store';
 
 const historySelectedStore = useHistorySelectedStore();
 const { selectedId } = storeToRefs(historySelectedStore);
