+++ src/modules/histories/utils/history-categorizer.ts
