+++ src/modules/stuffs/components/StuffDetailItemList/StuffDetailItemList.vue
@@ -5,25 +5,25 @@ import { NIL as NIL_UUID } from 'uuid';
 import { onBeforeMount, ref, watchEffect } from 'vue';
 import { useMutation, useQueryClient } from 'vue-query';
 
-import { addNewItem } from '@common/apis/beliemeApis';
-import { stuffKeys } from '@common/apis/queryKeys';
-import buildAlertModal from '@common/components/AlertModal/utils/alertModalBuilder';
+import { addNewItem } from '@common/apis/belieme-apis';
+import { stuffKeys } from '@common/apis/query-keys';
+import buildAlertModal from '@common/components/AlertModal/utils/alert-modal-builder';
 import BasicModal from '@common/components/BasicModal/BasicModal.vue';
 import type BaseError from '@common/errors/BaseError';
 import type ItemInfoOnly from '@common/models/ItemInfoOnly';
 import type StuffWithItems from '@common/models/StuffWithItems';
-import useDeptStore from '@common/stores/deptStore';
-import useModalStore from '@common/stores/modalStore';
+import useDeptStore from '@common/stores/dept-store';
+import useModalStore from '@common/stores/modal-store';
 
 import ItemListCell from '@^stuffs/components/StuffDetailItemListCell/StuffDetailItemListCell.vue';
 import {
   getStuffDetailQuery,
   getStuffListQuery,
   reloadStuffDataUsingCacheAndResponse,
-} from '@^stuffs/components/utils/utils';
-import useNewStuffInfo from '@^stuffs/stores/newStuffInfoStore';
-import useStuffDetailViewModeStore from '@^stuffs/stores/stuffDetailViewModeStore';
-import useStuffSelectedStore from '@^stuffs/stores/stuffSelectedStore';
+} from '@^stuffs/components/utils/stuff-query-utils';
+import useNewStuffInfo from '@^stuffs/stores/new-stuff-info-store';
+import useStuffDetailViewModeStore from '@^stuffs/stores/stuff-detail-view-mode-store';
+import useStuffSelectedStore from '@^stuffs/stores/stuff-selected-store';
 
 const MAX_ITEM_NUM = 50;
