+++ src/modules/stuffs/components/StuffPageOnEmpty/StuffPageOnEmpty.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
 import { storeToRefs } from 'pinia';
 
-import useUserStore from '@common/stores/userStore';
+import useUserStore from '@common/stores/user-store';
 
 import StuffDetail from '@^stuffs/components/StuffDetailSection/StuffDetailSection.vue';
-import useStuffDetailViewModeStore from '@^stuffs/stores/stuffDetailViewModeStore';
+import useStuffDetailViewModeStore from '@^stuffs/stores/stuff-detail-view-mode-store';
 
 const userStore = useUserStore();
 const { userMode } = storeToRefs(userStore);
