+++ src/modules/stuffs/stores/new-stuff-info-store.ts
